refactor: PresentationRequestSchema tests
