+++ tests/validation.test.ts
@@ -80,77 +80,6 @@ test('StoredCredentialSchema validation', async (t) => {
 });
 
 test('PresentationRequestSchema validation', async (t) => {
-  const spec = Spec(
-    {
-      signedData: Credential.Simple({ age: Field, name: Bytes32 }),
-      targetAge: Claim(Field),
-      targetName: Constant(Bytes32, Bytes32.fromString('Alice')),
-    },
-    ({ signedData, targetAge, targetName }) => ({
-      assert: Operation.and(
-        Operation.equals(Operation.property(signedData, 'age'), targetAge),
-        Operation.equals(Operation.property(signedData, 'name'), targetName)
-      ),
-      outputClaim: Operation.property(signedData, 'age'),
-    })
-  );
-
-  await t.test('validates no-context presentation request', () => {
-    let requestInitial = PresentationRequest.noContext(spec, {
-      targetAge: Field(18),
-    });
-    let serialized = JSON.parse(PresentationRequest.toJSON(requestInitial));
-
-    const result = PresentationRequestSchema.safeParse(serialized);
-    assert(
-      result.success,
-      'No-context presentation request should be valid: ' +
-        (result.success ? '' : JSON.stringify(result.error.issues, null, 2))
-    );
-  });
-
-  await t.test('validates zk-app presentation request', () => {
-    const request = PresentationRequest.zkApp(
-      spec,
-      {
-        targetAge: Field(18),
-      },
-      {
-        action: Field(123), // Mock method ID + args hash
-      }
-    );
-
-    const serialized = JSON.parse(PresentationRequest.toJSON(request));
-
-    const result = PresentationRequestSchema.safeParse(serialized);
-    assert(
-      result.success,
-      'ZkApp presentation request should be valid: ' +
-        (result.success ? '' : JSON.stringify(result.error.issues, null, 2))
-    );
-  });
-
-  await t.test('validates https presentation request', () => {
-    const request = PresentationRequest.https(
-      spec,
-      {
-        targetAge: Field(18),
-      },
-      {
-        action: 'POST /api/verify',
-      }
-    );
-
-    const serialized = JSON.parse(PresentationRequest.toJSON(request));
-
-    const result = PresentationRequestSchema.safeParse(serialized);
-    assert(
-      result.success,
-      'HTTPS presentation request should be valid: ' +
-        (result.success ? '' : JSON.stringify(result.error.issues, null, 2))
-    );
-  });
-
   await t.test('validates complex presentation request', () => {
     const InputData = { age: Field, name: Bytes32 };
     const NestedInputData = { person: InputData, points: Field };
